requirements.txt
# Berlin AI Talent Radar — Python Dependencies
# Install: pip install -r requirements.txt

# Core
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
requests>=2.28.0

# Collectors
beautifulsoup4>=4.12.0

# Embeddings & RAG (Stage 4)
openai>=1.0.0

# Analytics & Dashboard (Stages 5-6)
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0

# Development (optional)
# pytest>=7.0.0
# python-dotenv>=1.0.0
